+++ sample/KebabManager.Web/App_Start/RazorGeneratorMvcStart.cs
@@ -3,9 +3,9 @@ using System.Web.Mvc;
 using System.Web.WebPages;
 using RazorGenerator.Mvc;
 
-[assembly: WebActivatorEx.PostApplicationStartMethod(typeof(KebabManager.Web.App_Start.RazorGeneratorMvcStart), "Start")]
+[assembly: WebActivatorEx.PostApplicationStartMethod(typeof(KebabManager.Web.RazorGeneratorMvcStart), "Start")]
 
-namespace KebabManager.Web.App_Start {
+namespace KebabManager.Web {
     public static class RazorGeneratorMvcStart {
         public static void Start() {
             var engine = new PrecompiledMvcEngine(typeof(RazorGeneratorMvcStart).Assembly) {